backend/pyproject.toml
[tool.poetry]
name = "test_task_big_data"
version = "0.1.0"
description = "Тестовое задание в компанию \"BigData\""
authors = ["garmoshka <*>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
psycopg2 = "^2.9.3"
pytest-django = "^4.5.2"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
drf-spectacular = "^0.22.1"
gunicorn = "^20.1.0"
black = "^22.3.0"
factory-boy = "^3.2.1"
celery = "^5.2.7"
redis = "^4.3.3"

[tool.black]
exclude = "migrations"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
